databricks_apps/streamlit-chatbot-app/requirements.txt
mlflow>=2.21.2
streamlit==1.44.1
databricks-sdk
openai